+++ setup/calibration/__init__.py
@@ -1,0 +1,1 @@
+"""Internal calibration flows used by setup/calibrate_gantry.py."""
